consolidate SyncProvider and useSync files
