--- src/components/Sync/useSync.tsx
@@ -1,6 +1,0 @@
-import { useContext } from "react";
-import { SyncContext } from "./SyncProvider";
-
-const useSync = () => useContext(SyncContext);
-
-export default useSync;
